refactor: Config 패키지 명 수정
