+++ src/main/java/com/bidnamu/bidnamubackend/auth/config/CorsFilterConfig.java
@@ -1,4 +1,4 @@
-package com.bidnamu.bidnamubackend.auth.common;
+package com.bidnamu.bidnamubackend.auth.config;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
